Fix exec, missing condition

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 # Terminating a process with exec
 
 exec {'kill':
+  onlyif  => 'pgrep killmenow'
   command => 'pkill killmenow'
 }
